updated code
I have updated the code for the frame processing and graph generator to incorporate the speed for vehicles and to fix some bugs. Additionally, the new graph loader loads the data as a heterogeneous graph, which now includes all of the node types and node features (with some toggle parameters to turn some of the features off if necessary). I also made a new jupyter notebook (task_experiments) that includes some experiments using these graphs with a Heterogenous GNN from pytorch geometric for a couple of different downstream tasks. In these case the node features are processed using a simple MLP (to be updated to a better architecture later). This does not yet include the images, but it does include the annotations that were extracted from the imaegs using shubham’s code. So far the performance for these tasks is not bad and the model seems to learn well.